Working Sentinel-2 diagnostics Vortex task
Significant modifications of s2m_command and vortex_kitchen have
been necessary.

- The -f option must be optional (no forcing involved)
- DBUG --uenv usage when a proper uenv is parsed (which should be the
  standard use)
- Some additionnal "if" options have been added. This could be avoided
  by a standardisation of vortex tasks.

+++ snowtools/scripts/post_processing/compute_sentinel2_diagnostics.py
@@ -25,13 +25,13 @@ def parse_command_line():
     parser = argparse.ArgumentParser(description=description)
     parser.add_argument('-b', '--datebegin', type=str, required=True,
                         help="First date covered by the simulation file, format YYYYMMDDHH.")
-    parser.add_argument('-e', '--dateend', type=str,
+    parser.add_argument('-e', '--dateend', type=str, default=None,
                         help="Last date covered by the simulation file, format YYYYMMDDHH.")
-    parser.add_argument('-x', '--xpid', type=str, required=True,
+    parser.add_argument('-x', '--xpid', type=str, default=None,
                         help="XPID of the simulation format XP_NAME@username")
     parser.add_argument('-m', '--members', type=int, default=None,
                         help="Number of members associated to the experiment")
-    parser.add_argument('-w', '--workdir', type=str, required=True,
+    parser.add_argument('-w', '--workdir', type=str, default=None,
                         help="Working directory")
     parser.add_argument('-k', '--mask', type=str,
                         help="Absolute path to the mask file")
@@ -76,9 +76,49 @@ def lcscd(data, threshold=.2):
     """
 
     data = xr.where(data > threshold, True, False)
-    cumulative = data.cumsum(dim='time') - data.cumsum(dim='time').where(data.values == 0).ffill(dim='time').fillna(0)
-    scd = (cumulative.max(dim = 'time')).rename('scd_concurent')
-    mod = (cumulative.argmax(dim = 'time') + 1).rename('mod')
+    # The goal of the following is to compute an array containing the number of concurent snow covered days :
+    # * 0 for snow free days
+    # * else : number of days since the begining of the current snow covered period
+    # Example :
+    # data = np.array([True,False,True,True,True,False,True,True])
+    # the expected output is :
+    # cumul = [1, 0, 1, 2, 3, 0, 1, 1]
+    # then :
+    # * scd = 3 (max(cumul))
+    # * mod = 6 (max(cumul) on position 4 --> first snow free day on position 5 = 6th day since the begining)
+    # * sod = 3 (3th day since the begining of the season)
+
+    try:
+        # 1. Smart solution to use on local computer or sxcen
+        # WARNING : ffill does not work on Meteo-France's HPC dur to a dependency to the "bottleneck" module,
+        # which is not currently installed on belneos/taranis
+        # data.cumsum(dim='time') = array([1, 1, 2, 3, 4, 4, 5, 6]) (numpy syntax : data.cumsum())
+        # data.cumsum(dim='time').where(data.values == 0) = array([1, 4]) (numpy : data.cumsum()[np.where(data==0)])
+        # data.cumsum(dim='time').where(data.values == 0).ffill(dim='time') = array([nan, 1, 1, 1, 1, 4, 4, 4])
+        # data.cumsum(dim='time').where(data.values == 0).ffill(dim='time').fillna(0) = [0, 1, 1, 1, 1, 4, 4, 4]
+        # cumul = array([1, 0, 1, 2, 3, 0, 1, 2])
+        cumul = data.cumsum(dim='time') - data.cumsum(dim='time').where(data.values == 0).ffill(dim='time').fillna(0)
+    except ModuleNotFoundError:
+        # This workaround is 10 times slower than the (better) solution above but it works on Meteo-France's HPC and
+        # it returns the same result.
+        def scd_periods_np(array):
+            # This method is appplied to each 1D array (implicit loop over the domain's pixels) of the
+            # data.cumsum(dim='time').values numpy array.
+            # In the previous example, array would be array([1, 1, 2, 3, 4, 4, 5, 6])
+            out = np.zeros(np.shape(array))  # array([0, 0, 0, 0, 0, 0, 0, 0])
+            out[0] = array[0]  # Initialisation of the first date --> array([1, 0, 0, 0, 0, 0, 0, 0])
+            for idx in range(1, len(array)):
+                if array[idx] == array[idx - 1]:
+                    out[idx] = 0  # snow free date --> reset the counter (array([1, 0,...]))
+                else:
+                    out[idx] = out[idx - 1] + 1  # increment the current concurent snow cover period duration
+                    # array([1, 0, 1,...]) ; array([1, 0, 1, 2,...]) ; array([1, 0, 1, 2, 3,...])
+            # out = array([1, 0, 1, 2, 3, 0, 1, 2])
+            return out
+        cumul = data.copy()
+        cumul.values = np.apply_along_axis(scd_periods_np, 0, data.cumsum(dim='time').values)
+    scd = (cumul.max(dim = 'time')).rename('scd_concurent')
+    mod = (cumul.argmax(dim = 'time') + 2).rename('mod')
     sod = (mod - scd).rename('sod')
     sd = data.where(data, np.nan).count(dim = 'time').rename('sd')
     return xr.merge([scd, mod, sod, sd])
@@ -138,31 +178,34 @@ if __name__ == '__main__':
     pro       = args.pro
     geometry  = 'GrandesRousses250m'
 
-    os.chdir(workdir)
+    if workdir is not None:
+        os.chdir(workdir)
 
     if mask is not None:
         # Get mask file
         # import shutil
         # shutil.copyfile(mask, 'mask.nc')
-        os.symlink(mask, 'mask.nc')
+        if not os.path.exists('mask.nc'):  # TODO remplacer le lien par sécurité ?
+            os.symlink(mask, 'mask.nc')
         mask = True
 
-    try:
-        # Retrieve PRO files with Vortex
-        from snowtools.scripts.extract.vortex import vortexIO
-        vortexIO.get_pro(datebegin, dateend, xpid, geometry, members=members)
-    except (ImportError, ModuleNotFoundError):
-        # TODO : Le code actuel ne gère qu'un unique fichier PRO
-        # TODO : il reste à gérer les simulations d'ensemble avec 1 PRO/membre
-        # shutil.copyfile(pro, 'PRO.nc')
-        os.symlink(pro, 'PRO.nc')
+    if xpid is not None:
+        try:
+            # Retrieve PRO files with Vortex
+            from snowtools.scripts.extract.vortex import vortexIO
+            vortexIO.get_pro(datebegin, dateend, xpid, geometry, members=members)
+        except (ImportError, ModuleNotFoundError):
+            # TODO : Le code actuel ne gère qu'un unique fichier PRO
+            # TODO : il reste à gérer les simulations d'ensemble avec 1 PRO/membre
+            # shutil.copyfile(pro, 'PRO.nc')
+            os.symlink(pro, 'PRO.nc')
 
     if members is None:
         subdir = ''
         diag(subdir, datebegin, mask=mask)
         os.remove('PRO.nc')
     else:
-        for member in range(1, members + 1):
+        for member in range(members):
             print(f'Member {member}')
             subdir = f'mb{member:03d}'
             diag(subdir, datebegin, mask=mask)
